refactor: [1단계-문자열 계산기] Domain 로직에서 Controller 분리
