+++ src/test/java/calculatortest/CalculatorTest.java
@@ -1,9 +1,10 @@
 
 package calculatortest;
 
-import calculator.domain.Calculator;
+import calculator.domain.Calculate;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -12,15 +13,15 @@ import java.io.ByteArrayInputStream;
 
 public class CalculatorTest {
     //given
-    Calculator calculator = new Calculator();
+    Calculate calculate = new Calculate();
 
-    @ParameterizedTest
-    @DisplayName("calculator")
-    @ValueSource(strings = {"4 + 3 * 2 / 1"})
-    public void calculator(String string) {
-        ByteArrayInputStream input = new ByteArrayInputStream(string.getBytes());
-        System.setIn(input);
-        Assertions.assertThat(calculator.calculate()).isEqualTo(14);
+    @Test
+    @DisplayName("calculate")
+    public void calculate() {
+        // given
+
+        // when, then
+        Assertions.assertThat(calculate.calculate()).isEqualTo(14);
     }
 
     @ParameterizedTest
